+++ messages/de/models_Page.php
@@ -1,12 +1,11 @@
 <?php
-
-return [
-    'Navigation' => 'Navigation',
-    'Only visible for admins' => 'Nur für Admins sichtbar',
-    'Open in new window' => 'In neuem Fenster öffnen',
-    'Url shortcut' => 'URL-Verknüpfung',
-    'View' => 'Ansicht',
-    'page' => 'Seite',
-    'Abstract' => '',
-    'Page' => '',
-];
+return array (
+  'Abstract' => 'Beschreibung',
+  'Navigation' => 'Navigation',
+  'Only visible for admins' => 'Nur für Admins sichtbar',
+  'Open in new window' => 'In neuem Fenster öffnen',
+  'Page' => 'Seite',
+  'Url shortcut' => 'URL-Verknüpfung',
+  'View' => 'Ansicht',
+  'page' => 'Seite',
+);