+++ messages/de/models_Snippet.php
@@ -2,5 +2,5 @@
 return array (
   'Content' => 'Inhalt',
   'Sidebar' => 'Seitenleiste (Sidebar)',
-  'snippet' => 'Schnipsel',
+  'snippet' => 'Snippet',
 );