+++ messages/de/modules_template_controller_SnippedTemplateAdminController.php
@@ -1,4 +1,4 @@
 <?php
 return array (
-  'Here you can manage your snipped layouts. Snippet layouts are templates, which can be included into sidebars.' => 'Hier kannst du deine Schnipsel-Layouts verwalten. Schnipsel-Layouts sind Vorlagen, die in Sidebars eingebunden werden können.',
+  'Here you can manage your snipped layouts. Snippet layouts are templates, which can be included into sidebars.' => 'Hier kannst du deine Snippet-Layouts verwalten. Snippet-Layouts sind Vorlagen, die in Sidebars eingebunden werden können.',
 );