+++ messages/es/controllers_TemplateController.php
@@ -1,8 +1,8 @@
 <?php
 return array (
-  '<strong>Edit</strong> {type} element' => '',
+  '<strong>Edit</strong> {type} element' => '<strong> Editar </strong> {type} elemento',
   'Access denied!' => 'Acceso denegado!',
-  'Empty content elements cannot be delted!' => '',
-  'Invalid request data!' => '',
-  'You are not allowed to delete default content!' => '',
+  'Empty content elements cannot be delted!' => '¡Los elementos de contenido vacíos no se pueden eliminar!',
+  'Invalid request data!' => 'Datos de solicitud no válidos',
+  'You are not allowed to delete default content!' => '¡No puedes eliminar contenido predeterminado!',
 );