+++ messages/es/models_RichtextContent.php
@@ -1,7 +1,6 @@
 <?php
-
-return [
-    'Empty HumHub Richtext' => '',
-    'Empty Richtext' => '',
-    'Empty Text' => '',
-];
+return array (
+  'Empty HumHub Richtext' => 'Texto enriquecido de HumHub vacío',
+  'Empty Richtext' => 'Texto enriquecido vacío',
+  'Empty Text' => 'Texto vacío',
+);