+++ messages/es/models_TemplateElement.php
@@ -1,7 +1,7 @@
 <?php
 return array (
   'Label' => 'Etiqueta',
-  'Placeholder name' => '',
-  'The element name must contain at least two characters without spaces or special signs except \'_\'' => '',
-  'The given element name is already in use for this template.' => '',
+  'Placeholder name' => 'Nombre del marcador de posición',
+  'The element name must contain at least two characters without spaces or special signs except \'_\'' => 'El nombre del elemento debe contener al menos dos caracteres sin espacios ni signos especiales excepto \'_\'',
+  'The given element name is already in use for this template.' => 'El nombre del elemento dado ya está en uso para esta plantilla.',
 );