+++ messages/es/modules_template_views_admin_editSource.php
@@ -1,9 +1,8 @@
 <?php
-
-return [
-    'Add Element' => '',
-    'Edit All' => '',
-    'Edit template \'{templateName}\'' => '',
-    'Here you can edit the source of your template by defining the template layout and adding content elements. Each element can be assigned with a default content and additional definitions.' => '',
-    'You haven\'t saved your last changes yet. Do you want to leave without saving?' => '',
-];
+return array (
+  'Add Element' => 'Agregar elemento',
+  'Edit All' => 'Editar todo',
+  'Edit template \'{templateName}\'' => 'Editar plantilla \'{templateName}\'',
+  'Here you can edit the source of your template by defining the template layout and adding content elements. Each element can be assigned with a default content and additional definitions.' => 'Aquí puede editar la fuente de su plantilla definiendo el diseño de la plantilla y agregando elementos de contenido. A cada elemento se le puede asignar un contenido predeterminado y definiciones adicionales.',
+  'You haven\'t saved your last changes yet. Do you want to leave without saving?' => 'Aún no has guardado tus últimos cambios. ¿Quieres irte sin guardar?',
+);