+++ messages/es/views_admin_add.php
@@ -1,7 +1,7 @@
 <?php
 return array (
-  'Add new {pageType}' => '',
-  'Create new template' => '',
-  'Edit template' => '',
+  'Add new {pageType}' => 'Agregar nuevo {pageType}',
+  'Create new template' => 'Crear nueva plantilla',
+  'Edit template' => 'Editar plantilla',
   'Settings' => 'Ajustes',
 );