+++ messages/es/widgets_views_editMultipleElements.php
@@ -1,7 +1,7 @@
 <?php
 return array (
   'Empty' => 'Vacío',
-  'Inline' => '',
-  'Multiple' => '',
-  'This template does not contain any elements yet.' => '',
+  'Inline' => 'En línea',
+  'Multiple' => 'Múltiple',
+  'This template does not contain any elements yet.' => 'Esta plantilla aún no contiene ningún elemento.',
 );