+++ messages/he/models_Page.php
@@ -1,12 +1,11 @@
 <?php
-
-return [
-    'Abstract' => '',
-    'Navigation' => '',
-    'Only visible for admins' => '',
-    'Open in new window' => '',
-    'Page' => '',
-    'Url shortcut' => '',
-    'View' => '',
-    'page' => '',
-];
+return array (
+  'Abstract' => '',
+  'Navigation' => '',
+  'Only visible for admins' => '',
+  'Open in new window' => 'פתח בחלון חדש',
+  'Page' => '',
+  'Url shortcut' => '',
+  'View' => '',
+  'page' => '',
+);