+++ messages/hu/models_RichtextContent.php
@@ -1,7 +1,6 @@
 <?php
-
-return [
-    'Empty Richtext' => 'Üres Richtext',
-    'Empty Text' => 'Üres szöveg',
-    'Empty HumHub Richtext' => '',
-];
+return array (
+  'Empty HumHub Richtext' => 'Üres HumHub Richtext',
+  'Empty Richtext' => 'Üres Richtext',
+  'Empty Text' => 'Üres szöveg',
+);