+++ messages/hu/views_common_edit.php
@@ -1,12 +1,11 @@
 <?php
-
-return [
-    'By setting an url shortcut value, you can create a better readable url for your page. If <b>URL Rewriting</b> is enabled on your site, the value \'mypage\' will result in an url \'www.example.de/p/mypage\'.' => 'Az URL shortcut érték beállításával jobban olvasható URL-címet hozhatsz létre az oldalad számára. Ha az <b>URL átírása</b> engedélyezve van a weboldaladon, akkor a "mypage" érték a következő url-t eredményezi: "www.example.de/p/mypage"',
-    'Configuration' => 'Konfiguráció',
-    'Delete' => 'Törlés',
-    'Inline Editor' => 'Inline szerkesztő',
-    'e.g. http://www.example.de' => 'pl.: http://www.valami.hu/',
-    'Here you can configure the general settings of your {pageLabel}.' => '',
-    'The abstract will be used as stream entry content to promote the actual page. 
-                        If no abstract is given or the page is only visible for admins, no stream entry will be created.' => '',
-];
+return array (
+  'By setting an url shortcut value, you can create a better readable url for your page. If <b>URL Rewriting</b> is enabled on your site, the value \'mypage\' will result in an url \'www.example.de/p/mypage\'.' => 'Az URL shortcut érték beállításával jobban olvasható URL-címet hozhatsz létre az oldalad számára. Ha az <b>URL átírása</b> engedélyezve van a weboldaladon, akkor a "mypage" érték a következő url-t eredményezi: "www.example.de/p/mypage"',
+  'Configuration' => 'Konfiguráció',
+  'Delete' => 'Törlés',
+  'Here you can configure the general settings of your {pageLabel}.' => 'Itt konfigurálhatja a {pageLabel} általános beállításait.',
+  'Inline Editor' => 'Inline szerkesztő',
+  'The abstract will be used as stream entry content to promote the actual page. 
+                        If no abstract is given or the page is only visible for admins, no stream entry will be created.' => 'Az összefoglaló az adatfolyam bejegyzésének tartalmaként lesz felhasználva az aktuális oldal népszerűsítésére. Ha nem ad hozzá összefoglalót, vagy az oldal csak az adminok számára látható, akkor nem jön létre adatfolyam bejegyzés.',
+  'e.g. http://www.example.de' => 'pl.: http://www.valami.hu/',
+);