+++ messages/hu/views_common_list.php
@@ -1,5 +1,4 @@
 <?php
-
-return [
-    'This page lists all available custom content entries.' => '',
-];
+return array (
+  'This page lists all available custom content entries.' => 'Ez az oldal felsorolja az összes rendelkezésre álló egyéni tartalmi bejegyzést.',
+);