+++ messages/sq/modules_template_controller_OwnerContentController.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    '<strong>Confirm</strong> container item deletion' => '',
+    '<strong>Confirm</strong> content deletion' => '',
+    '<strong>Confirm</strong> element deletion' => '',
+];