+++ messages/sq/modules_template_views_element_addItemChooseTemplate.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'Choose a template' => '',
+    'Template' => '',
+];