+++ modules/template/widgets/PageConfigurationButton.php
@@ -9,10 +9,7 @@
 namespace humhub\modules\custom_pages\modules\template\widgets;
 
 use Yii;
-use humhub\modules\custom_pages\models\ContainerPage;
-use humhub\modules\custom_pages\models\Page;
 use humhub\modules\custom_pages\modules\template\models\PagePermission;
-use humhub\modules\custom_pages\modules\template\models\TemplateInstance;
 
 /**
  * User Administration Menu
@@ -21,32 +18,42 @@ use humhub\modules\custom_pages\modules\template\models\TemplateInstance;
  */
 class PageConfigurationButton extends \humhub\components\Widget
 {
-    public $canEdit;
-    public $editMode;
+    /**
+     * @var integer
+     */
     public $pageId;
-    public $templateInstance;
+
+    /**
+     * @var string
+     */
+    public $target = '_blank';
+
+    /**
+     * @var string
+     */
+    public $btnClass = 'btn btn-primary btn-xs';
+
+    /**
+     * @var string
+     */
+    public $btnStyles = 'margin-bottom:5px';
 
     public function run()
     {
-        $editMode = Yii::$app->request->get('editMode');
-        $pageId = Yii::$app->request->get('id');
+        if (!PagePermission::canEdit()) {
+            return;
+        }
 
-        $space = (isset(Yii::$app->controller->contentContainer)) ? Yii::$app->controller->contentContainer : null;
+        $pageId = isset($this->pageId) ? $this->pageId : Yii::$app->request->get('id');
 
-        $sguid = ($space != null) ? $space->guid : null;
-        $canEdit = PagePermission::canEdit();
-
-        $ownerModel = ($space != null) ? ContainerPage::class : Page::class;
-
-
-        $templateInstance = TemplateInstance::findOne(['object_model' => $ownerModel ,'object_id' => $pageId]);
+        $contentContainer = isset(Yii::$app->controller->contentContainer) ? Yii::$app->controller->contentContainer : null;
 
         return $this->render('pageConfigurationButton', [
-            'canEdit' => $canEdit,
-            'sguid' => $sguid,
-            'editMode' => $editMode,
             'pageId' => $pageId,
-            'templateInstance' => $templateInstance
+            'contentContainer' => $contentContainer,
+            'target' => $this->target,
+            'btnClass' => $this->btnClass,
+            'btnStyles' => $this->btnStyles,
         ]);
     }