+++ docs/CHANGELOG.md
@@ -1,9 +1,30 @@
 Changelog
 =========
 
-1.2.1 (Unreleased)
+1.4.0 (Unreleased)
 -------------------------
 - Enh #163: Attachments for HTML pages
+- Enh: HumHub min. version increased to 1.8
+
+
+1.3.1 (February 23, 2021)
+-------------------------
+- Fix: Stream Channel was set for AdminOnly pages
+- Enh: Add info regarding "Admin Only" pages without Stream support
+
+
+1.3.0 (February 22, 2021)
+-------------------------
+- Fix #88: Text alignment resets when reloading ckeditor input
+- Chng: Removed ckeditor show more toolbar item due to content filter issues
+- Chng: Updated ckeditor to 4.16.0
+- Fix #168: Add edit button to markdown pages
+
+
+1.2.1 (February 05, 2021)
+-------------------------
+- Fix #166 Only system admin can access custom pages space settings
+
 
 1.2.0 (December 12, 2020)
 -------------------------